+++ main.go
@@ -10,31 +10,20 @@ import (
 	"os"
 	"os/signal"
 	"runtime"
-	"sync"
 	"syscall"
 	"time"
 
 	"github.com/bugsnag/bugsnag-go"
 
-	"github.com/rudderlabs/rudder-server/services/diagnostics"
+	"github.com/rudderlabs/rudder-server/services/apptype"
 
 	"github.com/rudderlabs/rudder-server/admin"
 	"github.com/rudderlabs/rudder-server/app"
 	"github.com/rudderlabs/rudder-server/config"
 	backendconfig "github.com/rudderlabs/rudder-server/config/backend-config"
-	"github.com/rudderlabs/rudder-server/gateway"
-	"github.com/rudderlabs/rudder-server/jobsdb"
-	"github.com/rudderlabs/rudder-server/processor"
-	ratelimiter "github.com/rudderlabs/rudder-server/rate-limiter"
 	"github.com/rudderlabs/rudder-server/router"
-	"github.com/rudderlabs/rudder-server/router/batchrouter"
 	"github.com/rudderlabs/rudder-server/rruntime"
-	"github.com/rudderlabs/rudder-server/services/db"
-	destinationdebugger "github.com/rudderlabs/rudder-server/services/destination-debugger"
-	sourcedebugger "github.com/rudderlabs/rudder-server/services/source-debugger"
 	"github.com/rudderlabs/rudder-server/services/stats"
-	"github.com/rudderlabs/rudder-server/services/validators"
-	"github.com/rudderlabs/rudder-server/utils"
 	"github.com/rudderlabs/rudder-server/utils/logger"
 	"github.com/rudderlabs/rudder-server/utils/misc"
 	"github.com/rudderlabs/rudder-server/utils/types"
@@ -45,37 +34,43 @@ import (
 )
 
 var (
-	application                      app.Interface
-	warehouseMode                    string
-	maxProcess                       int
-	gwDBRetention, routerDBRetention time.Duration
-	enableProcessor, enableRouter    bool
-	enabledDestinations              []backendconfig.DestinationT
-	configSubscriberLock             sync.RWMutex
-	objectStorageDestinations        []string
-	warehouseDestinations            []string
-	moduleLoadLock                   sync.Mutex
-	routerLoaded                     bool
-	processorLoaded                  bool
-	enableSuppressUserFeature        bool
-	pkgLogger                        logger.LoggerI
-	Diagnostics                      diagnostics.DiagnosticsI = diagnostics.Diagnostics
+	application               app.Interface
+	warehouseMode             string
+	enableSuppressUserFeature bool
+	pkgLogger                 logger.LoggerI
+	appType                   AppTypeHandler
 )
 
 var version = "Not an official release. Get the latest release from the github repo."
 var major, minor, commit, buildDate, builtBy, gitURL, patch string
 
-func loadConfig() {
-	maxProcess = config.GetInt("maxProcess", 12)
-	gwDBRetention = config.GetDuration("gwDBRetentionInHr", 0) * time.Hour
-	routerDBRetention = config.GetDuration("routerDBRetention", 0)
-	enableProcessor = config.GetBool("enableProcessor", true)
-	enableRouter = config.GetBool("enableRouter", true)
-	objectStorageDestinations = []string{"S3", "GCS", "AZURE_BLOB", "MINIO", "DIGITAL_OCEAN_SPACES"}
-	warehouseDestinations = []string{"RS", "BQ", "SNOWFLAKE", "POSTGRES", "CLICKHOUSE"}
-	warehouseMode = config.GetString("Warehouse.mode", "embedded")
-	// Enable suppress user feature. false by default
-	enableSuppressUserFeature = config.GetBool("Gateway.enableSuppressUserFeature", false)
+const (
+	gatewayAppType   = "GATEWAY"
+	processorAppType = "PROCESSOR"
+	monolithAppType  = "MONOLITH"
+)
+
+//AppTypeHandler to be implemented by different app type objects.
+type AppTypeHandler interface {
+	GetAppType() string
+	HandleRecovery(*app.Options)
+	StartRudderCore(*app.Options)
+}
+
+func getAppType(application app.Interface, appType string) AppTypeHandler {
+	var handler AppTypeHandler
+	switch appType {
+	case gatewayAppType:
+		handler = &apptype.GatewayAppType{App: application, VersionHandler: versionHandler}
+	case processorAppType:
+		handler = &apptype.ProcessorAppType{App: application, VersionHandler: versionHandler}
+	case monolithAppType:
+		handler = &apptype.MonolithAppType{App: application, VersionHandler: versionHandler}
+	default:
+		panic(errors.New("invalid app type"))
+	}
+
+	return handler
 }
 
 // Test Function
@@ -91,42 +86,9 @@ func readIOforResume(router router.HandleT) {
 	}
 }
 
-// Gets the config from config backend and extracts enabled writekeys
-func monitorDestRouters(routerDB, batchRouterDB *jobsdb.HandleT) {
-	ch := make(chan utils.DataEvent)
-	backendconfig.Subscribe(ch, backendconfig.TopicBackendConfig)
-	dstToRouter := make(map[string]*router.HandleT)
-	dstToBatchRouter := make(map[string]*batchrouter.HandleT)
-	// dstToWhRouter := make(map[string]*warehouse.HandleT)
-
-	for {
-		config := <-ch
-		sources := config.Data.(backendconfig.SourcesT)
-		enabledDestinations := make(map[string]bool)
-		for _, source := range sources.Sources {
-			for _, destination := range source.Destinations {
-				enabledDestinations[destination.DestinationDefinition.Name] = true
-				//For batch router destinations
-				if misc.Contains(objectStorageDestinations, destination.DestinationDefinition.Name) || misc.Contains(warehouseDestinations, destination.DestinationDefinition.Name) {
-					_, ok := dstToBatchRouter[destination.DestinationDefinition.Name]
-					if !ok {
-						pkgLogger.Info("Starting a new Batch Destination Router ", destination.DestinationDefinition.Name)
-						var brt batchrouter.HandleT
-						brt.Setup(batchRouterDB, destination.DestinationDefinition.Name)
-						dstToBatchRouter[destination.DestinationDefinition.Name] = &brt
-					}
-				} else {
-					_, ok := dstToRouter[destination.DestinationDefinition.Name]
-					if !ok {
-						pkgLogger.Info("Starting a new Destination ", destination.DestinationDefinition.Name)
-						var router router.HandleT
-						router.Setup(routerDB, destination.DestinationDefinition.Name)
-						dstToRouter[destination.DestinationDefinition.Name] = &router
-					}
-				}
-			}
-		}
-	}
+func loadConfig() {
+	warehouseMode = config.GetString("Warehouse.mode", "embedded")
+	enableSuppressUserFeature = config.GetBool("Gateway.enableSuppressUserFeature", false)
 }
 
 func init() {
@@ -154,114 +116,8 @@ func startWarehouseService() {
 	warehouse.Start()
 }
 
-func startRudderCore(clearDB *bool, normalMode bool, degradedMode bool) {
-	pkgLogger.Info("Main starting")
-
-	if !validators.ValidateEnv() {
-		panic(errors.New("Failed to start rudder-server"))
-	}
-	validators.InitializeEnv()
-
-	// Check if there is a probable inconsistent state of Data
-	if diagnostics.EnableServerStartMetric {
-		Diagnostics.Track(diagnostics.ServerStart, map[string]interface{}{
-			diagnostics.ServerStart: fmt.Sprint(time.Unix(misc.AppStartTime, 0)),
-		})
-	}
-
-	//Reload Config
-	loadConfig()
-
-	var gatewayDB jobsdb.HandleT
-	var routerDB jobsdb.HandleT
-	var batchRouterDB jobsdb.HandleT
-	var procErrorDB jobsdb.HandleT
-
-	runtime.GOMAXPROCS(maxProcess)
-	pkgLogger.Info("Clearing DB ", *clearDB)
-
-	destinationdebugger.Setup()
-	sourcedebugger.Setup()
-
-	migrationMode := application.Options().MigrationMode
-	gatewayDB.Setup(*clearDB, "gw", gwDBRetention, migrationMode, false)
-
-	enableGateway := true
-
-	if application.Features().Migrator != nil {
-		if migrationMode == db.IMPORT || migrationMode == db.EXPORT || migrationMode == db.IMPORT_EXPORT {
-			startProcessorFunc := func() {
-				clearDBBool := false
-				StartProcessor(&clearDBBool, migrationMode, enableProcessor, &gatewayDB, &routerDB, &batchRouterDB, &procErrorDB)
-			}
-			startRouterFunc := func() {
-				StartRouter(enableRouter, &routerDB, &batchRouterDB)
-			}
-			enableRouter = false
-			enableProcessor = false
-			enableGateway = (migrationMode != db.EXPORT)
-			application.Features().Migrator.Setup(&gatewayDB, &routerDB, &batchRouterDB, startProcessorFunc, startRouterFunc)
-		}
-	}
-
-	StartProcessor(clearDB, migrationMode, enableProcessor, &gatewayDB, &routerDB, &batchRouterDB, &procErrorDB)
-	StartRouter(enableRouter, &routerDB, &batchRouterDB)
-
-	if enableGateway {
-		var gateway gateway.HandleT
-		var rateLimiter ratelimiter.HandleT
-
-		rateLimiter.SetUp()
-		gateway.Setup(application, backendconfig.DefaultBackendConfig, &gatewayDB, &rateLimiter, clearDB, versionHandler)
-		gateway.StartWebHandler()
-	}
-	//go readIOforResume(router) //keeping it as input from IO, to be replaced by UI
-}
-
-//NOTE: StartProcessor and StartRouter should be called in order and from the same thread.
-//StartProcessor sets up router, batch router, proc error DBs, which router also depends on.
-//enableRouter will be true only if enableProcessor is true. This is a must.
-//If otherwise, server may crash because setup of router and batch router DB is not called.
-
-//StartProcessor atomically starts processor process if not already started
-func StartProcessor(clearDB *bool, migrationMode string, enableProcessor bool, gatewayDB, routerDB, batchRouterDB *jobsdb.HandleT, procErrorDB *jobsdb.HandleT) {
-	moduleLoadLock.Lock()
-	defer moduleLoadLock.Unlock()
-
-	if processorLoaded {
-		return
-	}
-
-	if enableProcessor {
-		//setting up router, batch router, proc error DBs only if processor is enabled.
-		routerDB.Setup(*clearDB, "rt", routerDBRetention, migrationMode, true)
-		batchRouterDB.Setup(*clearDB, "batch_rt", routerDBRetention, migrationMode, true)
-		procErrorDB.Setup(*clearDB, "proc_error", routerDBRetention, migrationMode, false)
-
-		var processor = processor.NewProcessor()
-		processor.Setup(backendconfig.DefaultBackendConfig, gatewayDB, routerDB, batchRouterDB, procErrorDB)
-		processor.Start()
-
-		processorLoaded = true
-	}
-}
-
-//StartRouter atomically starts router process if not already started
-func StartRouter(enableRouter bool, routerDB, batchRouterDB *jobsdb.HandleT) {
-	moduleLoadLock.Lock()
-	defer moduleLoadLock.Unlock()
-
-	if routerLoaded {
-		return
-	}
-
-	if enableRouter {
-		go monitorDestRouters(routerDB, batchRouterDB)
-		routerLoaded = true
-	}
-}
-
 func canStartServer() bool {
+	pkgLogger.Info("warehousemode ", warehouseMode)
 	return warehouseMode == config.EmbeddedMode || warehouseMode == config.OffMode
 }
 
@@ -270,6 +126,20 @@ func canStartWarehouse() bool {
 }
 
 func main() {
+	options := app.LoadOptions()
+	if options.VersionFlag {
+		printVersion()
+		return
+	}
+
+	application = app.New(options)
+
+	//TODO initEnterpriseFeatures based on apptype?
+	//application & backend setup should be done before starting any new goroutines.
+	application.Setup()
+
+	appType = getAppType(application, config.GetEnv("APP_TYPE", ""))
+
 	version := versionInfo()
 
 	bugsnag.Configure(bugsnag.Configuration{
@@ -278,7 +148,7 @@ func main() {
 		// The import paths for the Go packages containing your source files
 		ProjectPackages: []string{"main", "github.com/rudderlabs/rudder-server"},
 		// more configuration options
-		AppType:      "rudder-server",
+		AppType:      appType.GetAppType(),
 		AppVersion:   version["Version"].(string),
 		PanicHandler: func() {},
 	})
@@ -298,19 +168,6 @@ func main() {
 
 	//Creating Stats Client should be done right after setting up logger and before setting up other modules.
 	stats.Setup()
-
-	options := app.LoadOptions()
-	if options.VersionFlag {
-		printVersion()
-		return
-	}
-
-	application = app.New(options)
-
-	http.HandleFunc("/version", versionHandler)
-
-	//application & backend setup should be done before starting any new goroutines.
-	application.Setup()
 
 	var pollRegulations bool
 	if enableSuppressUserFeature {
@@ -343,19 +200,19 @@ func main() {
 
 	misc.AppStartTime = time.Now().Unix()
 	if canStartServer() {
-		db.HandleRecovery(options.NormalMode, options.DegradedMode, options.MigrationMode, misc.AppStartTime)
-
+		appType.HandleRecovery(options)
 		rruntime.Go(func() {
-			startRudderCore(&options.ClearDB, options.NormalMode, options.DegradedMode)
+			appType.StartRudderCore(options)
 		})
 	}
 
-	// initialize warehouse service after core to handle non-normal recovery modes
+	//TODO: Handle warehouse later
+	/*// initialize warehouse service after core to handle non-normal recovery modes
 	if canStartWarehouse() {
 		rruntime.Go(func() {
 			startWarehouseService()
 		})
-	}
+	}*/
 
 	rruntime.Go(admin.StartServer)
